requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic[email]
mysql-connector-python
python-decouple
python-jose
passlib
python-multipart